requirements.txt
absl-py==1.4.0
accelerate==0.16.0
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
alembic==1.11.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
autogluon==0.8.2
autogluon.common==0.8.2
autogluon.core==0.8.2
autogluon.features==0.8.2
autogluon.multimodal==0.8.2
autogluon.tabular==0.8.2
autogluon.timeseries==0.8.2
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
blis==0.7.10
boto3==1.28.16
botocore==1.31.16
cachetools==5.3.1
catalogue==2.0.9
catboost==1.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
colorful==0.5.5
confection==0.1.0
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
databricks-cli==0.17.7
datasets==2.14.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
docker==6.1.3
entrypoints==0.4
evaluate==0.3.0
exceptiongroup==1.1.2
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.2
Flask==2.3.2
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
gluonts==0.13.2
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gpustat==1.1
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.50.0
gunicorn==20.1.0
h11==0.14.0
huggingface-hub==0.16.4
hyperopt==0.2.7
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
lightgbm==3.3.5
lightning-utilities==0.9.0
llvmlite==0.40.1
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mlflow==2.5.0
mlforecast==0.7.3
model-index==0.1.11
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
networkx==3.1
nlpaug==1.1.11
nltk==3.8.1
nptyping==2.4.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==12.535.77
oauthlib==3.2.2
omegaconf==2.2.3
opencensus==0.11.2
opencensus-context==0.1.3
opendatalab==0.0.9
openmim==0.3.9
ordered-set==4.1.0
outcome==1.2.0
packaging==23.1
pandas==1.5.3
pathy==0.10.2
patsy==0.5.3
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.15.0
preshed==3.0.8
prometheus-client==0.17.1
protobuf==3.20.2
psutil==5.9.5
psycopg2-binary==2.9.7
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytorch-metric-learning==1.7.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
querystring-parser==1.2.4
ray==2.3.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.5.1
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.11.1
selenium==4.11.2
sentencepiece==0.1.99
seqeval==1.2.2
shap==0.42.1
six==1.16.0
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.19
sqlparse==0.4.4
srsly==2.4.7
statsforecast==1.4.0
statsmodels==0.14.0
streamlit==1.25.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2
text-unidecode==1.3
thinc==8.1.10
threadpoolctl==3.2.0
tifffile==2023.7.18
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.3.2
tqdm==4.65.0
transformers==4.26.1
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
urllib3==1.26.16
validators==0.20.0
virtualenv==20.24.2
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.1
Werkzeug==2.3.6
window-ops==0.0.14
wsproto==1.2.0
xgboost==1.7.6
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
